requirements.txt
asgiref
autopep8
Django==2.1.*
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-import-export
djangorestframework
djangorestframework-simplejwt==4.3.0
graphene-django